Fix final issues

+++ app/auth/signout/page.tsx
@@ -3,10 +3,12 @@ import { useRouter } from "next/navigation";
 
 export default function SignOut() {
 
+  const router = useRouter();
+
   // Remove all local storage variables including token
-  localStorage.clear();
+  if (typeof window !== "undefined")
+    localStorage.clear();
 
   // Return back to landing page
-  const router = useRouter();
   router.push("/");
 }
